+++ SpaceMaster App/app/src/androidTest/java/com/example/myapplication4/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.SpaceMaster;
+package com.example.myapplication4;
 
 import android.content.Context;
