+++ SpaceMaster App/app/src/main/AndroidManifest.xml
@@ -13,13 +13,13 @@
         android:theme="@style/Theme.MyApplication4"
         tools:targetApi="31">
         <activity
-            android:name=".java.SignUpActivity"
+            android:name=".SignUpActivity"
             android:exported="false" />
         <activity
-            android:name=".java.MainActivity2"
+            android:name=".MainActivity2"
             android:exported="false" />
         <activity
-            android:name=".java.SignInActivity"
+            android:name=".SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
