--- SpaceMaster App/app/src/main/res/drawable/baseline_account_balance_24.xml
@@ -1,9 +1,0 @@
-<vector android:height="24dp" android:tint="#D0D2D0"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M4,10h3v7h-3z"/>
-    <path android:fillColor="@android:color/white" android:pathData="M10.5,10h3v7h-3z"/>
-    <path android:fillColor="@android:color/white" android:pathData="M2,19h20v3h-20z"/>
-    <path android:fillColor="@android:color/white" android:pathData="M17,10h3v7h-3z"/>
-    <path android:fillColor="@android:color/white" android:pathData="M12,1l-10,5l0,2l20,0l0,-2z"/>
-</vector>
