--- SpaceMaster App/app/src/main/res/drawable/baseline_space_dashboard_24.xml
@@ -1,5 +1,0 @@
-<vector android:height="24dp" android:tint="#C0FCDD"
-    android:viewportHeight="24" android:viewportWidth="24"
-    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="@android:color/white" android:pathData="M11,21H5c-1.1,0 -2,-0.9 -2,-2V5c0,-1.1 0.9,-2 2,-2h6V21zM13,21h6c1.1,0 2,-0.9 2,-2v-7h-8V21zM21,10V5c0,-1.1 -0.9,-2 -2,-2h-6v7H21z"/>
-</vector>
