+++ SpaceMaster App/app/src/main/res/drawable/edittext_rounded_bg.xml
@@ -2,6 +2,6 @@
     <solid android:color="@color/white" /> <!-- Replace with your desired background color -->
     <corners android:radius="10dp" /> <!-- Adjust the radius as per your preference -->
     <stroke
-        android:color="@color/blue"
+        android:color="@color/gray"
         android:width="1dp" /> <!-- Replace with your desired stroke color and width -->
 </shape>
