+++ SpaceMaster App/app/src/main/res/layout/activity_main2.xml
@@ -4,132 +4,73 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:background="@drawable/background"
-    tools:context=".java.MainActivity2">
+    android:backgroundTint="@color/black"
+    tools:context=".MainActivity2">
 
     <TextView
         android:id="@+id/textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="mainactivity2"
-        android:visibility="gone"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.46"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.852" />
-
-    <ImageView
-        android:id="@+id/image_areas"
-        android:layout_width="413dp"
-        android:layout_height="259dp"
-        android:scaleType="fitCenter"
-        android:src="@drawable/p2_ship_mask"
-        android:visibility="visible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.498" />
-
-    <ImageView
-        android:id="@+id/image_1"
-        android:layout_width="408dp"
-        android:layout_height="252dp"
-        android:alpha="0.4"
-        android:scaleType="fitCenter"
-        android:src="@drawable/p2_ship_default_1"
-        android:visibility="invisible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="1.0"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.502" />
-
-    <ImageView
-        android:id="@+id/image_2"
-        android:layout_width="412dp"
-        android:layout_height="255dp"
-        android:alpha="0.4"
-        android:scaleType="fitCenter"
-        android:src="@drawable/p2_ship_default_2"
-        android:visibility="visible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.497"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
     <ImageView
+        android:id="@+id/image_areas"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        android:scaleType="fitCenter"
+        android:src="@drawable/p2_ship_mask"
+        android:visibility="visible"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent" />
+
+    <ImageView
         android:id="@+id/image"
-        android:layout_width="412dp"
-        android:layout_height="256dp"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
         android:scaleType="fitCenter"
         android:src="@drawable/p2_ship_default"
         android:visibility="visible"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="@+id/image_1"
-        app:layout_constraintStart_toStartOf="@+id/image_1"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.495" />
+        tools:ignore="MissingConstraints" />
 
+    <ImageView
+        android:alpha="0.4"
+        android:id="@+id/image_1"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        android:scaleType="fitCenter"
+        android:src="@drawable/p2_ship_default_1"
+        android:visibility="invisible"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:ignore="MissingConstraints" />
+    <ImageView
+        android:alpha="0.4"
+        android:id="@+id/image_2"
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        android:scaleType="fitCenter"
+        android:src="@drawable/p2_ship_default_2"
+        android:visibility="invisible"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        tools:ignore="MissingConstraints" />
     <Button
         android:id="@+id/supa2button"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Button"
 
-        android:visibility="gone"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
-
-    <nl.joery.timerangepicker.TimeRangePicker
-        android:theme="@style/timePicker"
-        android:id="@+id/picker"
-        android:layout_width="230dp"
-        android:layout_height="213dp"
-
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.969"
-        app:trp_endTime="10:00"
-        app:trp_maxDurationMinutes="600"
-        app:trp_minDurationMinutes="30"
-        app:trp_sliderWidth="12dp"
-        app:trp_startTime="8:00"
-        app:trp_thumbIconEnd="@drawable/ic_launcher_background"
-        app:trp_thumbIconStart="@drawable/ic_launcher_background" />
-
-    <DatePicker
-        android:theme="@style/CalenderViewCustom"
-        android:dateTextAppearance="@style/CalenderViewDateCustomText"
-        android:weekDayTextAppearance="@style/CalenderViewWeekCustomText"
-        android:unfocusedMonthDateColor="@color/white"
-        android:selectedWeekBackgroundColor="@color/white"
-        android:weekSeparatorLineColor="@color/white"
-        android:focusedMonthDateColor="@color/white"
-        android:weekNumberColor="@color/white"
-        android:textColorPrimary="@color/white"
-        android:id="@+id/datePicker"
-        android:layout_width="306dp"
-        android:layout_height="188dp"
-        android:backgroundTint="#FFFFFF"
-        android:calendarViewShown="false"
-        android:datePickerMode="spinner"
-        android:headerBackground="#FFFFFF"
-        android:textColor="#FFFFFF"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.495"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.029" />
-
-
 </androidx.constraintlayout.widget.ConstraintLayout>
