--- SpaceMaster App/app/src/main/res/values/style.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <style name="CalenderViewCustom" parent="Theme.AppCompat">
-        <item name="colorAccent">@color/red</item>
-        <item name="colorPrimary">@color/white</item>
-    </style>
-
-    <style name="CalenderViewDateCustomText" parent="android:TextAppearance.DeviceDefault.Small">
-        <item name="android:textColor">@color/white</item>
-        <item name="android:weekNumberColor">@color/red</item>
-    </style>
-
-    <style name="CalenderViewWeekCustomText" parent="android:TextAppearance.DeviceDefault.Small">
-        <item name="android:textColor">@color/white</item>
-    </style>
-    <style name="timePicker" parent="Theme.AppCompat">
-        <item name="colorAccent">@color/blue</item>
-        <item name="colorPrimary">@color/blue</item>
-        <item name="android:textColor">@color/white</item>
-    </style>
-        <!-- Base application theme. -->
-</resources>
