+++ SpaceMaster App/app/src/test/java/com/example/myapplication4/ExampleUnitTest.java
@@ -1,4 +1,4 @@
-package com.example.SpaceMaster;
+package com.example.myapplication4;
 
 import org.junit.Test;
