+++ SpaceMaster App/build.gradle
@@ -1,17 +1,9 @@
 buildscript {
-    ext {
-        kotlin_version = '1.9.0-Beta'
-    }
     dependencies {
         classpath 'com.google.gms:google-services:4.3.15'
-        classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
-    }
-    repositories {
-        mavenCentral()
     }
 }// Top-level build file where you can add configuration options common to all sub-projects/modules.
 plugins {
     id 'com.android.application' version '8.0.1' apply false
     id 'com.android.library' version '8.0.1' apply false
-    id 'org.jetbrains.kotlin.android' version '1.9.0-Beta' apply false
 }
